+++ src/app/Pages/companyDetails/CompanyDetails.module.css
