+++ src/app/Pages/companyDetails/CompanyDetails.tsx
