+++ src/app/Pages/industries/Industries.module.css
