+++ src/app/Pages/industries/Industries.tsx
